\name{continue.ddp}
\alias{continue.ddp}
\docType{package}
\title{
Resume MCMC run
}
\description{
Resume MCMC run from an object of class ddp.
}
\usage{
\method{continue}{ddp}(previous,...)
}
\arguments{
\item{previous}{An object of class ddp}
\item{...}{Arguments to be passed to method}
}
\value{
An object of class ddp.
}
