\name{continue.ddpcomp}
\alias{continue.ddpcomp}
\docType{package}
\title{
Resume MCMC run
}
\description{
Resume MCMC run from an object of class ddpcomp.
}
\usage{
\method{continue}{ddpcomp}(previous,...)
}
\arguments{
\item{previous}{An object of class ddpcomp}
\item{...}{Arguments to be passed to method}
}
\value{
An object of class ddpcomp.
}
