\name{continue.dpm}
\alias{continue.dpm}
\docType{package}
\title{
Resume MCMC run
}
\description{
Resume MCMC run from an object of class dpm.
}
\usage{
\method{continue}{dpm}(previous,...)
}
\arguments{
\item{previous}{An object of class dpm}
\item{...}{Arguments to be passed to method}
}
\value{
An object of class dpm.
}
