\name{continue.dpmcomp}
\alias{continue.dpmcomp}
\docType{package}
\title{
Resume MCMC run
}
\description{
Resume MCMC run from an object of class dpmcomp.
}
\usage{
\method{continue}{dpmcomp}(previous,...)
}
\arguments{
\item{previous}{An object of class dpmcomp}
\item{...}{Arguments to be passed to method}
}
\value{
An object of class dpmcomp.
}
