\name{plot.ddp}
\alias{plot.ddp}
\docType{package}
\title{
plot estimated log hazard ratio functions from an object of class ddp.
}
\description{
plot estimated log hazard ratio functions with credible intervals from an object of class ddp.
}
\usage{
\method{plot}{ddp}(x,simultaneous=FALSE, exp=FALSE, ...)
}
\arguments{
\item{x}{Output an object of class ddp}
\item{simultaneous}{Plot simultaneous credible intervals or not. The default is FALSE.}
\item{exp}{Plot hazard ratio (TRUE) or log hazard ratio (FALSE). The default is FALSE.}
\item{...}{Arguments to be passed to method}
}
\value{
plot estimated log hazard ratio functions from an object of class ddp.
}
