\name{plot.ddpcomp}
\alias{plot.ddpcomp}
\docType{package}
\title{
plot estimated log subdistribution hazard ratio functions for cause 1 from an object of class ddpcomp.
}
\description{
plot estimated log subdistribution hazard ratio functions with credible intervals for cause 1 from an object of class ddpcomp.
}
\usage{
\method{plot}{ddpcomp}(x,simultaneous=FALSE,exp=FALSE,...)
}
\arguments{
\item{x}{Output an object of class ddpcomp}
\item{simultaneous}{Plot simultaneous credible intervals or not. The default is FALSE.}
\item{exp}{Plot hazard ratio (TRUE) or log hazard ratio (FALSE). The default is FALSE.}
\item{...}{Arguments to be passed to method}
}
\value{
plot estimated log subdistribution hazard ratio functions for cause 1 from an object of class ddpcomp.
}
