\name{plot.dpm}
\alias{plot.dpm}
\docType{package}
\title{
plot estimated survival/density/hazard functions from an object of class dpm.
}
\description{
plot estimated survival/density/hazard functions with credible intervals from an object of class dpm.
}
\usage{
\method{plot}{dpm}(x,simultaneous=FALSE,...)
}
\arguments{
\item{x}{Output an object of class dpm}
\item{simultaneous}{Plot simultaneous credible intervals or not. The default is FALSE.}
\item{...}{Arguments to be passed to method}
}
\value{
plot estimated survival/density/hazard functions from an object of class dpm.
}
