\name{plot.dpmcomp}
\alias{plot.dpmcomp}
\docType{package}
\title{
plot estimated cumulative incidence/ subdistribution density/ subdistribution hazard functions from an object of class dpmcomp.
}
\description{
plot estimated cumulative incidence/ subdistribution density/ subdistribution hazard functions with credible intervals from an object of class dpmcomp.
}
\usage{
\method{plot}{dpmcomp}(x,simultaneous=FALSE,...)
}
\arguments{
\item{x}{Output an object of class dpmcomp}
\item{simultaneous}{Plot simultaneous credible intervals or not. The default is FALSE.}
\item{...}{Arguments to be passed to method}
}
\value{
plot estimated cumulative incidence/ subdistribution density/ subdistribution hazard functions from an object of class dpm.
}
