\name{plot.predddp}
\alias{plot.predddp}
\docType{package}
\title{
plot estimated survival/density/hazard functions from an object of class predddp.
}
\description{
plot estimated survival/density/hazard functions with credible intervals from an object of class predddp.
}
\usage{
\method{plot}{predddp}(x,...)
}
\arguments{
\item{x}{Output an object of class predddp}
\item{...}{Arguments to be passed to method}
}
\value{
plot estimated survival/density/hazard functions from an object of class predddp.
}
