\name{plot.predddpcomp}
\alias{plot.predddpcomp}
\docType{package}
\title{
plot estimated cumulative incidence function/ subdistribution density/ subdistribution hazard functions of cause 1 from an object of class predddpcomp.
}
\description{
plot estimated cumulative incidence function/ subdistribution density/ subdistribution hazard functions with credible intervals of cause 1 from an object of class predddpcomp.
}
\usage{
\method{plot}{predddpcomp}(x,...)
}
\arguments{
\item{x}{Output an object of class predddpcomp}
\item{...}{Arguments to be passed to method}
}
\value{
plot estimated cumulative incidence function/ subdistribution density/ subdistribution hazard functions of cause 1 from an object of class predddpcomp.
}
