\name{summary.ddp}
\alias{summary.ddp}
\docType{package}
\title{
generate summary of the dpweib output with ddp class.
}
\description{
generate estimated log hazard ratio and corresponding credible intervals of the dpweib output with ddp class at 4 time points (1/4, 1/2, 3/4 and 1 of high.pct).
}
\usage{
\method{summary}{ddp}(object,...)
}
\arguments{
\item{object}{Output from dpweib of ddp class}
\item{...}{Arguments to be passed to method}
}
\value{
an object of class summary.ddp
}
