\name{summary.ddpcomp}
\alias{summary.ddpcomp}
\docType{package}
\title{
generate summary of the dpweib output with ddpcomp class.
}
\description{
generate estimated log subdistribution hazard ratio and the corresponding credible intervals of the dpweib output with ddpcomp class at 4 time points (1/4, 1/2, 3/4 and 1 of high.pct).
}
\usage{
\method{summary}{ddpcomp}(object,...)
}
\arguments{
\item{object}{Output from dpweib of ddpcomp class}
\item{...}{Arguments to be passed to method}
}
\value{
an object of class summary.ddpcomp
}
