\name{summary.dpm}
\alias{summary.dpm}
\docType{package}
\title{
generate summary of the dpweib output with dpm class.
}
\description{
generate estimated survival and corresponding credible intervals of the dpweib output with dpm class at 4 time points(1/4, 1/2, 3/4 and 1 of high.pct).
}
\usage{
\method{summary}{dpm}(object,...)
}
\arguments{
\item{object}{Output from dpweib of class dpm}
\item{...}{Arguments to be passed to method}
}
\value{
an object of class summary.dpm
}
