\name{summary.dpmcomp}
\alias{summary.dpmcomp}
\docType{package}
\title{
generate summary of the dpweib output with dpmcomp class.
}
\description{
generate estimated cumulative incidence functions and corresponding credible intervals of the dpweib output with dpmcomp class at 4 time points (1/4, 1/2, 3/4 and 1 of high.pct).
}
\usage{
\method{summary}{dpmcomp}(object,...)
}
\arguments{
\item{object}{Output from dpweib of class dpmcomp}
\item{...}{Arguments to be passed to method}
}
\value{
an object of class summary.dpmcomp
}
