\name{survmedian}
\alias{survmedian}
\docType{data}
\alias{Median of Survivals}
\title{The median of the surivival functions generated by LIO prior}
\description{
  This data set gives the median of 20000 random survival functions generated by our LIO prior on a 0 to 10 scale. This data set is primarily used for determining scale parameter when heavy end-of-study censoring is present.
}
\source{
Shi,Y. and Martens,M. and Banerjee,A. and Laud,P. (2017) Low Information Omnibus Priors for Dirichlet Process Mixture Models(Manuscript)
}
